requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
Pillow>=9.5.0
opencv-python>=4.7.0
onnx>=1.14.0
onnxruntime>=1.15.0
matplotlib>=3.7.1
ultralytics>=8.0.0
